Added check whether symbolic link exists

+++ link.sh
@@ -2,5 +2,7 @@
 set -e
 
 for file in $(ls -ap | grep '^\.' | grep -v '/$'); do
-    ln -sv $(pwd)/${file} ~/${file}
+    if [[ ! -L ~/${file} ]]  && [[ ! -e ~/${file} ]] ; then
+        ln -sv $(pwd)/${file} ~/${file}
+    fi
 done
